% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogdata_preprocess.R
\name{cogdata_preprocess}
\alias{cogdata_preprocess}
\title{Read WearIT Survey data ----}
\usage{
cogdata_preprocess(data_path)
}
\arguments{
\item{data_path}{data path containing cogtest results export}
}
\description{
Read WearIT Survey data ----
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
