% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_fns.R
\name{get_data_fns}
\alias{get_data_fns}
\title{Read WearIT Survey data ----
\code{get_data_fns} takes path to a data folder as input and returns list of all files}
\usage{
get_data_fns(data_path = NA)
}
\arguments{
\item{data_path}{path to data folder}
}
\description{
Read WearIT Survey data ----
\code{get_data_fns} takes path to a data folder as input and returns list of all files
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
