% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_csvs.R
\name{list_csvs}
\alias{list_csvs}
\title{Read WearIT Survey data ----
\code{list_csvs} takes path to a data folder as input and returns list of all files}
\usage{
list_csvs(data_path = NA)
}
\arguments{
\item{data_path}{path to data folder}
}
\description{
Read WearIT Survey data ----
\code{list_csvs} takes path to a data folder as input and returns list of all files
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
