% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_csvs.R
\name{merge_csvs}
\alias{merge_csvs}
\title{Read WearIT Survey data ----}
\usage{
merge_csvs(fns = NA, use_labels = TRUE, drop_deprecated = TRUE)
}
\value{
Returns stacked dataframe
}
\description{
Read WearIT Survey data ----
}
\details{
Must specify which type of file to stack (key files, nonkey files, target files, or all files)
}
