% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data_collection.R
\name{merge_data_collection}
\alias{merge_data_collection}
\title{Read WearIT Survey data ----}
\usage{
merge_data_collection(fns = NA)
}
\value{
Returns stacked dataframe
}
\description{
Read WearIT Survey data ----
}
\details{
Must specify which type of file to stack (key files, nonkey files, target files, or all files)
}
