% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsecheck_cogtask_data.R
\name{parsecheck_cogtask_data}
\alias{parsecheck_cogtask_data}
\title{Read WearIT Survey data ----}
\usage{
parsecheck_cogtask_data(original_data, unnested_data)
}
\arguments{
\item{original_data}{original, pre-processed data}

\item{unnested_data}{result of `unnest_cogtask_data()`}
}
\description{
Read WearIT Survey data ----
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
