% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_cogtask_data.R
\name{preprocess_cogtask_data}
\alias{preprocess_cogtask_data}
\title{Read WearIT Survey data ----}
\usage{
preprocess_cogtask_data(data_path)
}
\arguments{
\item{data_path}{data path containing cogtest results export}
}
\description{
Read WearIT Survey data ----
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
