% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_surveydata.R
\name{read_surveydata}
\alias{read_surveydata}
\title{Read WearIT Survey data ----}
\usage{
read_surveydata(fn, col_names = NA, output = c("key_table", "tidy_data"))
}
\arguments{
\item{fn}{path to .csv data file}

\item{col_names}{???}

\item{output}{type of table to be produced. Options include "key_table" 
(produces list of question numbers, questions, and variable names/labels), "tidy_data" 
(produces list including all data), or c("key_table", "tidy_data") to produce both.}
}
\description{
Read WearIT Survey data ----
}
\details{
None
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
