% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wearit_surveydata.R
\name{read_wearit_surveydata}
\alias{read_wearit_surveydata}
\title{Read WearIT Survey data ----}
\usage{
read_wearit_surveydata(fn, output = c("key_table", "tidy_data"))
}
\description{
Read WearIT Survey data ----
}
