% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Read WearIT Survey data ----}
\usage{
run_pipeline(
  data_path,
  config_path,
  use_labels = T,
  silent = TRUE,
  drop_deprecated = TRUE
)
}
\description{
Read WearIT Survey data ----
}
