% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline_cogdata_rev.R
\name{run_pipeline_cogdata_rev}
\alias{run_pipeline_cogdata_rev}
\title{Read WearIT Survey data ----}
\usage{
run_pipeline_cogdata_rev(data_path, config_path)
}
\description{
Read WearIT Survey data ----
}
