% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline_survey.R
\name{run_pipeline_survey}
\alias{run_pipeline_survey}
\title{Read WearIT Survey data ----}
\usage{
run_pipeline_survey(
  data_path,
  config_path,
  use_labels = TRUE,
  drop_deprecated = TRUE
)
}
\description{
Read WearIT Survey data ----
}
