% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_study_datapipeline.R
\name{run_study_datapipeline}
\alias{run_study_datapipeline}
\title{Read WearIT Survey data ----}
\usage{
run_study_datapipeline(data_path, config_path)
}
\description{
Read WearIT Survey data ----
}
