% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_pipeline_cogtasks.R
\name{study_pipeline_cogtasks}
\alias{study_pipeline_cogtasks}
\title{Read WearIT Survey data ----}
\usage{
study_pipeline_cogtasks(data_path, config_path)
}
\description{
Read WearIT Survey data ----
}
