% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_cogtask_data.R
\name{unnest_cogtask_data}
\alias{unnest_cogtask_data}
\title{Read WearIT Survey data ----}
\usage{
unnest_cogtask_data(.data)
}
\arguments{
\item{data}{data object returned from `preprocess_cogtask_data()`}
}
\description{
\code{unnest_cogtask_data} unnests M2C2 Cogtask JSON data from WearIT platform
}
\author{
Nelson Roque, \email{Nelson.Roque@ucf.edu}
}
