% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_collection.R
\name{create_data_collection}
\alias{create_data_collection}
\title{Read WearIT Survey data ----}
\usage{
create_data_collection(data_path = NA, types = c())
}
\description{
Read WearIT Survey data ----
}
